+++ src/characters/PlasticBag.ts
@@ -5,7 +5,7 @@ class PlasticBag extends Obstacle {
   protected readonly _imageFilename: string = "plastic-bag.png";
   protected readonly _foodValue: number = 0;
   protected readonly _damage: number = 8;
-  protected readonly _speed: number = 0.1;
+  protected readonly _speed: number = 0;
   protected readonly _stomachImpact: number = 20;
   protected readonly _points: number = -20;
 }