+++ src/characters/abstract/Obstacle.ts
@@ -1,3 +1,4 @@
+import Game from "../../Game";
 import { takeDamage } from "../../features/turtleMonitor/turtleReducers";
 import store from "../../store";
 import NonMain from "./NonMain";
@@ -9,8 +10,12 @@ abstract class Obstacle extends NonMain {
   }
 
   swim(): void {
-    this._x -= this._speed;
-    super.swim();
+    const speed = Game.instance.level.currentSpeed;
+    this._x -= speed;
+    const randomVeriticalDirection = Math.round(Math.random());
+    const randomHorizontalDirection = Math.round(Math.random());
+    this._y += randomVeriticalDirection ? speed : -speed;
+    this._x += randomHorizontalDirection ? speed : -speed;
   }
 }