+++ src/levels/Level2.ts
@@ -6,8 +6,10 @@ class Level2 extends Level {
   protected readonly _initialCharacters: LevelCharacter[] = [
     { type: "plasticBag", amount: 10 },
     { type: "neptuneGrass", amount: 4 },
+    { type: "shrimp", amount: 30 },
   ];
   protected readonly _benthicOffsetY: number = 600;
+  protected readonly _currentSpeed: number = 0.5;
 }
 
 export default Level2;